requirements.txt
flask
pydantic
defusedxml
google-cloud-storage
celery[redis]
google-cloud-pubsub